+++ internal/dto/movies/movie_request.go
@@ -1,17 +1,17 @@
 package movies
 
 type CreateMovieRequest struct {
-	Title           string `json:"title" binding:"required,min=1,max=255"`
-	Description     string `json:"description" binding:"omitempty,max=2000"`
-	PosterImage     string `json:"poster_image" binding:"omitempty,url"`
-	DurationMinutes int    `json:"duration_minutes" binding:"required,gte=1,lte=600"`
-	GenreIDs        []uint `json:"genre_ids" binding:"required,min=1,dive,gt=0"`
+	Title           string   `json:"title" binding:"required,min=1,max=255"`
+	Description     string   `json:"description" binding:"omitempty,max=2000"`
+	PosterImage     string   `json:"poster_image" binding:"omitempty,url"`
+	DurationMinutes int      `json:"duration_minutes" binding:"required,gte=1,lte=600"`
+	GenreIDs        []uint64 `json:"genre_ids" binding:"required,min=1,dive,gt=0"`
 }
 
 type UpdateMovieRequest struct {
-	Title           string `json:"title" binding:"omitempty,min=1,max=255"`
-	Description     string `json:"description" binding:"omitempty,max=2000"`
-	PosterImage     string `json:"poster_image" binding:"omitempty,url"`
-	DurationMinutes int    `json:"duration_minutes" binding:"omitempty,gte=1,lte=600"`
-	GenreIDs        []uint `json:"genre_ids" binding:"omitempty,min=1,dive,gt=0"`
+	Title           string   `json:"title" binding:"omitempty,min=1,max=255"`
+	Description     string   `json:"description" binding:"omitempty,max=2000"`
+	PosterImage     string   `json:"poster_image" binding:"omitempty,url"`
+	DurationMinutes int      `json:"duration_minutes" binding:"omitempty,gte=1,lte=600"`
+	GenreIDs        []uint64 `json:"genre_ids" binding:"omitempty,min=1,dive,gt=0"`
 }
